refactor(settings): move isTimeFormat into server-safe @/lib/time-format (#453)
PR #452 closed #427 by importing `isTimeFormat` from
`src/hooks/useUserSettings.ts` into the server-side
`src/app/api/settings/route.ts` PUT handler. That works (Next's bundler
tree-shakes the hook's React/client code) but routes a server-only
validator through a `"use client"` module — exactly the boundary
ambiguity that motivated `VALID_DATE_FORMATS` living in
`src/lib/format-date.ts` (no `"use client"`).

This change extracts `TTimeFormat`, `VALID_TIME_FORMATS`, and
`isTimeFormat` into a new server-safe module `src/lib/time-format.ts`
that mirrors `format-date.ts` in shape and intent. Every consumer
(route handler, hook, settings page, display section, CalendarProvider,
fixtures, types) is migrated in place rather than retained via a
re-export — no module re-imports the symbols from `@/hooks/*` anymore.

Behaviour is unchanged at every boundary:
- The PUT handler still rejects unknown `timeFormat` with the same 400
  status and the same `"Invalid timeFormat. Must be one of: 12h, 24h"`
  body.
- All hook semantics, settings-form rendering, and CalendarProvider
  derivations identical.

The `isTimeFormat` describe block in
`src/hooks/__tests__/useUserSettings.test.tsx` moves to a new
`src/lib/__tests__/time-format.test.ts` (test-first) so the
typeguard's tests live next to the typeguard.

Closes #453
