+++ src/components/providers/CalendarProvider.tsx
@@ -4,7 +4,7 @@ import type { CalendarsResponse } from "@/app/api/calendar/calendars/route";
 import { useProfileOptional } from "@/components/profiles/profile-context";
 import { useEventCacheVisibilitySweep } from "@/hooks/useEventCacheVisibilitySweep";
 import { useLocalStorage } from "@/hooks/useLocalStorage";
-import { type TTimeFormat, useUserSettings } from "@/hooks/useUserSettings";
+import { useUserSettings } from "@/hooks/useUserSettings";
 import {
   type HiddenEventCounts,
   ZERO_HIDDEN_COUNTS,
@@ -29,6 +29,7 @@ import {
 import { resolveTransitionDurationMs } from "@/lib/calendar/transition-speed";
 import type { GoogleCalendarEvent } from "@/lib/google-calendar";
 import { logger } from "@/lib/logger";
+import type { TTimeFormat } from "@/lib/time-format";
 import type {
   ICalendarInfo,
   IEvent,
